Updated TicketService, added validation when booking tickets, Exceptions and logs

+++ src/main/java/com/site/aviatrails/domain/tickets/Ticket.java
@@ -27,5 +27,5 @@ public class Ticket {
     private Integer numberOfTickets;
 
     @Column(name = "ticket_price")
-    private Long ticketPrice;
+    private Integer ticketPrice;
 }
